Getting index values from carousel change
