+++ client/src/components/main/RootRouter.jsx
@@ -7,11 +7,11 @@ import FollowingFeed from './FollowingFeed.jsx';
 import FollowingsPage from './FollowingsPage.jsx';
 import UserProfilePage from './UserProfilePage.jsx';
 import CurrentUserProfilePage from './CurrentUserProfilePage.jsx';
-import Timeline from './Timeline.jsx';
 import ProfileRouterPage from './ProfileRouterPage.jsx';
-import TestPage from './TestPage.jsx';
+import Timeline from './Timeline.jsx';
 import Demo from './Chart.jsx';
 import LandmarkPage from './LandmarkPage.jsx';
+import TestPage from './TestPage.jsx'
 
 const RootRouter = () => (
   <Switch>
@@ -22,9 +22,10 @@ const RootRouter = () => (
     <Route exact path='/post/:id' component={PostPage} />
     <Route exact path='/landmark/:id' component={LandmarkPage} />
     <Route exact path='/followings' component={FollowingsPage} />
-    <Route exact path='/testpage' component={TimeLine} />
+    <Route exact path='/testpage' component={TestPage} />
+    <Route exact path='/timeline' component={Timeline} />
   </Switch>
 );
 
-// <Route exact path='/timeline' component={Timeline} />
+
 export default RootRouter;
